Address style issues and move some functions around
