+++ include/commandreplyparser.h
@@ -34,6 +34,9 @@
 #include "tensorbase.h"
 #include "enums/cpp_tensor_type.h"
 
+using inner_map = std::unordered_map<std::string, std::string>;
+using parsed_map = std::unordered_map<std::string, inner_map>;
+
 namespace SmartRedis {
 
 namespace CommandReplyParser {
@@ -90,6 +93,51 @@ inline TensorType get_tensor_data_type(CommandReply& reply)
 
     return TENSOR_TYPE_MAP.at(std::string(reply[1].str(),
                                             reply[1].str_len()));
+}
+
+inline parsed_map parse_db_node_info(std::string info)
+{
+    /* Parse database node information from get_db_node_info()
+    *          into a nested unordered_map
+    *  Returns parsed_map containing the database node information
+    */
+    parsed_map info_map;
+
+    std::string delim = "\r\n";
+    std::string currKey = "";
+    size_t start = 0U;
+    size_t end = info.find(delim);
+
+    while (end != std::string::npos)
+    {
+        std::string line = info.substr(start, end-start);
+        start = end + delim.length();
+        end = info.find(delim, start);
+        if (line.length() == 0)
+            continue;
+        if (line[0] == '#')
+        {
+            currKey = line.substr(2);
+            if (info_map.find(currKey)==info_map.end())
+                info_map[currKey] = {};
+        }
+        else
+        {
+            std::size_t separatorIdx = line.find(':');
+            info_map[currKey][line.substr(0, separatorIdx)] =
+                                line.substr(separatorIdx+1);
+        }
+    }
+    std::string line = info.substr(start);
+    if (line.length() > 0)
+    {
+        std::size_t separatorIdx = line.find(':');
+        if (info_map.find(currKey)==info_map.end())
+                    info_map[currKey] = {};
+        info_map[currKey][line.substr(0, separatorIdx)] =
+                            line.substr(separatorIdx+1);
+    }
+    return info_map;
 }
 
 } //namespace CommandReplyParser
