+++ src/cpp/redis.cpp
@@ -124,9 +124,9 @@ bool Redis::key_exists(const std::string& key)
 bool Redis::is_addressable(const std::string& address,
                            const uint64_t& port)
 {
-    return this->_valid_nodes.find(address + ":"
+    return this->_address_node_map.find(address + ":"
                         + std::to_string(port))
-                        != this->_valid_nodes.end();
+                        != this->_address_node_map.end();
 }
 
 
@@ -329,7 +329,7 @@ CommandReply Redis::get_script(const std::string& key)
 
 inline void Redis::_connect(std::string address_port)
 {
-    this->_valid_nodes.insert({address_port, nullptr});
+    this->_address_node_map.insert({address_port, nullptr});
 
     int n_connection_trials = 10;
